increase the maximum size of query results to 2^64
Currently, we cannot select more than 2^32 rows from a table because we are limited by types of
variables containing the numbers of rows. This patch changes these types and sets new limits.

In classes which are being serialized and used in messaging, in order to be able to process queries
originating from older nodes, the top 32 bits of new integers are optional and stay at the end
of the class - if they're not present, we assume they equal 0.

Fixes #5101.

+++ cql3/statements/select_statement.hh
@@ -148,11 +148,11 @@ public:
     bool has_group_by() const { return _group_by_cell_indices && !_group_by_cell_indices->empty(); }
 
 protected:
-    uint32_t do_get_limit(const query_options& options, ::shared_ptr<term> limit) const;
-    uint32_t get_limit(const query_options& options) const {
+    uint64_t do_get_limit(const query_options& options, ::shared_ptr<term> limit) const;
+    uint64_t get_limit(const query_options& options) const {
         return do_get_limit(options, _limit);
     }
-    uint32_t get_per_partition_limit(const query_options& options) const {
+    uint64_t get_per_partition_limit(const query_options& options) const {
         return do_get_limit(options, _per_partition_limit);
     }
     bool needs_post_query_ordering() const;
@@ -290,7 +290,7 @@ private:
     future<::shared_ptr<cql_transport::messages::result_message::rows>>read_posting_list(
             service::storage_proxy& proxy,
             const query_options& options,
-            int32_t limit,
+            uint64_t limit,
             service::query_state& state,
             gc_clock::time_point now,
             db::timeout_clock::time_point timeout,
